+++ scheduler/send_emails_at_specific_time.py
@@ -1,77 +1,19 @@
-# email_scheduler.py
+#scheduler/send_emails_at_specific_time.py
 
-import logging
-import os
-from dotenv import load_dotenv
-from email_sender import send_email
-from excel_reader import read_data_from_excel
-from generate_email_address import generate_email_address
-import time
 import datetime
+import logging
+import time
+from data_utils.generate_email_address import generate_email_address
+from email_utils.email_sender import send_email
+from scheduler.email_scheduler import load_email_settings, read_email_template, read_excel_data
 
-# Set up logging
 logger = logging.getLogger(__name__)
 MAX_RETRIES = 3
 
-def load_email_settings():
-    load_dotenv()
-    logger.info("Loaded environment variables from .env file.")
-    return os.getenv('EMAIL_USERNAME'), os.getenv('EMAIL_PASSWORD')
-
-def read_email_template():
-    with open('email_assets/email_template.txt', 'r') as file:
-        email_template = file.read()
-    logger.info("Read email template.")
-    return email_template
-
-def read_excel_data():
-    excel_file = 'recruiters.xlsx'
-    sheet_name = 'Sheet1'
-    logger.info(f"Reading data from Excel file: '{excel_file}', sheet: '{sheet_name}'")
-    return read_data_from_excel(excel_file, sheet_name)
-
-def send_emails_now(batch_size=10):
-    sender_email, sender_password = load_email_settings()
-    email_template = read_email_template()
-    data = read_excel_data()
-
-    # Split the data into batches
-    for i in range(0, len(data), batch_size):
-        batch = data[i:i + batch_size]
-
-        for row in batch:
-            retries = 0
-            while retries < MAX_RETRIES:
-                try:
-                    first_name, last_name, email, company_name, designation = row
-                    recipient_emails = generate_email_address(first_name, last_name, email, company_name)
-                    if isinstance(recipient_emails, tuple):
-                        for recipient_email in recipient_emails:
-                            subject = f"[Aastha Shukla]: Exploring Full-Time SDE Roles at {company_name}"
-                            message = email_template.format(first_name=first_name, last_name=last_name, email=recipient_email,
-                                                             company_name=company_name, designation=designation if designation else "esteemed employee")
-                            send_email(sender_email, sender_password, recipient_email, subject, message, company_name)
-                            logger.info(f"Email sent successfully to {recipient_email}")
-                    elif recipient_emails:
-                        subject = f"[Aastha Shukla]: Exploring Full-Time SDE Roles at {company_name}"
-                        message = email_template.format(first_name=first_name, last_name=last_name, email=recipient_emails,
-                                                         company_name=company_name, designation=designation if designation else "esteemed employee")
-                        send_email(sender_email, sender_password, recipient_emails, subject, message, company_name)
-                        logger.info(f"Email sent successfully to {recipient_emails}")
-                    
-                    # If email sent successfully, break out of the retry loop
-                    break
-                except Exception as e:
-                    logger.error(f"Error sending email: {e}")
-                    retries += 1
-                    logger.info(f"Retrying... Retry attempt {retries}/{MAX_RETRIES}")
-                    time.sleep(10)  # Wait for a few seconds before retrying
-
-            # If maximum retries reached without success, log error
-            if retries == MAX_RETRIES:
-                logger.error("Max retries reached. Unable to send email.")
-
 def schedule_emails():
+    """
+    Prompts user to schedule emails for a specific time.
+    """
     specific_time = input("Enter the specific time to schedule the email (HH:MM): ")
     try:
         hour, minute = map(int, specific_time.split(':'))
@@ -83,6 +25,14 @@ def schedule_emails():
         print("Invalid time format. Please enter the time in HH:MM format.")
 
 def send_emails_at_specific_time(hour, minute, batch_size=10):
+    """
+    Sends emails at a specific time.
+
+    Args:
+        hour (int): Hour to send emails.
+        minute (int): Minute to send emails.
+        batch_size (int): Number of emails to send in each batch.
+    """
     sender_email, sender_password = load_email_settings()
     email_template = read_email_template()
     data = read_excel_data()
